@@ -60,76 +99,86 @@
   },
   "homepage": "https://rollupjs.org/",
   "optionalDependencies": {
-    "fsevents": "~2.3.2"
+    "fsevents": "~2.3.2",
+    "@rollup/rollup-darwin-arm64": "4.5.1",
+    "@rollup/rollup-android-arm64": "4.5.1",
+    "@rollup/rollup-win32-arm64-msvc": "4.5.1",
+    "@rollup/rollup-linux-arm64-gnu": "4.5.1",
+    "@rollup/rollup-linux-arm64-musl": "4.5.1",
+    "@rollup/rollup-android-arm-eabi": "4.5.1",
+    "@rollup/rollup-linux-arm-gnueabihf": "4.5.1",
+    "@rollup/rollup-win32-ia32-msvc": "4.5.1",
+    "@rollup/rollup-darwin-x64": "4.5.1",
+    "@rollup/rollup-win32-x64-msvc": "4.5.1",
+    "@rollup/rollup-linux-x64-gnu": "4.5.1",
+    "@rollup/rollup-linux-x64-musl": "4.5.1"
   },
   "devDependenciesComments": {
     "@rollup/plugin-typescript": "It appears that 11.1.3 breaks sourcemaps"
   },
   "devDependencies": {
-    "@codemirror/commands": "^6.2.5",
+    "@codemirror/commands": "^6.3.0",
     "@codemirror/lang-javascript": "^6.2.1",
-    "@codemirror/language": "^6.9.0",
-    "@codemirror/search": "^6.5.3",
-    "@codemirror/state": "^6.2.1",
-    "@codemirror/view": "^6.19.0",
+    "@codemirror/language": "^6.9.2",
+    "@codemirror/search": "^6.5.4",
+    "@codemirror/state": "^6.3.1",
+    "@codemirror/view": "^6.22.0",
     "@jridgewell/sourcemap-codec": "^1.4.15",
-    "@mermaid-js/mermaid-cli": "^10.4.0",
-    "@rollup/plugin-alias": "^5.0.0",
-    "@rollup/plugin-buble": "^1.0.2",
-    "@rollup/plugin-commonjs": "^25.0.4",
-    "@rollup/plugin-json": "^6.0.0",
-    "@rollup/plugin-node-resolve": "^15.2.1",
-    "@rollup/plugin-replace": "^5.0.2",
-    "@rollup/plugin-terser": "^0.4.3",
-    "@rollup/plugin-typescript": "11.1.2",
-    "@rollup/pluginutils": "^5.0.4",
-    "@types/estree": "1.0.1",
-    "@types/mocha": "^10.0.1",
-    "@types/node": "~14.18.61",
-    "@types/yargs-parser": "^21.0.0",
-    "@typescript-eslint/eslint-plugin": "^6.7.2",
-    "@typescript-eslint/parser": "^6.7.2",
+    "@mermaid-js/mermaid-cli": "^10.6.0",
+    "@napi-rs/cli": "^2.16.5",
+    "@rollup/plugin-alias": "^5.0.1",
+    "@rollup/plugin-buble": "^1.0.3",
+    "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-json": "^6.0.1",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "@rollup/plugin-replace": "^5.0.5",
+    "@rollup/plugin-terser": "^0.4.4",
+    "@rollup/plugin-typescript": "11.1.5",
+    "@rollup/pluginutils": "^5.0.5",
+    "@types/estree": "1.0.5",
+    "@types/mocha": "^10.0.4",
+    "@types/node": "18.0.0",
+    "@types/yargs-parser": "^21.0.3",
+    "@typescript-eslint/eslint-plugin": "^6.11.0",
+    "@typescript-eslint/parser": "^6.11.0",
     "@vue/eslint-config-prettier": "^8.0.0",
     "@vue/eslint-config-typescript": "^12.0.0",
-    "acorn": "^8.10.0",
+    "acorn": "^8.11.2",
     "acorn-import-assertions": "^1.9.0",
-    "acorn-jsx": "^5.3.2",
-    "acorn-walk": "^8.2.0",
     "buble": "^0.20.0",
     "builtin-modules": "^3.3.0",
     "chokidar": "^3.5.3",
     "colorette": "^2.0.20",
-    "concurrently": "^8.2.1",
-    "core-js": "^3.32.2",
+    "concurrently": "^8.2.2",
+    "core-js": "^3.33.2",
     "date-time": "^4.0.0",
     "es5-shim": "^4.6.7",
     "es6-shim": "^0.35.8",
-    "eslint": "^8.49.0",
+    "eslint": "^8.53.0",
     "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-import": "^2.28.1",
-    "eslint-plugin-prettier": "^5.0.0",
-    "eslint-plugin-unicorn": "^48.0.1",
-    "eslint-plugin-vue": "^9.17.0",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-prettier": "^5.0.1",
+    "eslint-plugin-unicorn": "^49.0.0",
+    "eslint-plugin-vue": "^9.18.1",
     "fixturify": "^3.0.0",
     "flru": "^1.0.2",
     "fs-extra": "^11.1.1",
     "github-api": "^3.4.0",
-    "hash.js": "^1.1.7",
     "husky": "^8.0.3",
-    "inquirer": "^9.2.11",
+    "inquirer": "^9.2.12",
     "is-reference": "^3.0.2",
-    "lint-staged": "^14.0.1",
+    "lint-staged": "^15.1.0",
     "locate-character": "^3.0.0",
-    "magic-string": "^0.30.3",
+    "magic-string": "^0.30.5",
     "mocha": "^10.2.0",
     "nyc": "^15.1.0",
-    "pinia": "^2.1.6",
-    "prettier": "^3.0.3",
+    "pinia": "^2.1.7",
+    "prettier": "^3.1.0",
     "pretty-bytes": "^6.1.1",
     "pretty-ms": "^8.0.0",
     "requirejs": "^2.3.6",
-    "rollup": "^3.29.2",
-    "rollup-plugin-license": "^3.1.0",
+    "rollup": "^4.4.0",
+    "rollup-plugin-license": "^3.2.0",
     "rollup-plugin-string": "^3.0.0",
     "rollup-plugin-thatworks": "^1.0.4",
     "semver": "^7.5.4",
@@ -138,26 +187,29 @@
     "source-map": "^0.7.4",
     "source-map-support": "^0.5.21",
     "systemjs": "^6.14.2",
-    "terser": "^5.19.4",
+    "terser": "^5.24.0",
     "tslib": "^2.6.2",
     "typescript": "^5.2.2",
-    "vite": "^4.4.9",
-    "vitepress": "^1.0.0-rc.14",
-    "vue": "^3.3.4",
+    "vite": "^5.0.0",
+    "vitepress": "^1.0.0-rc.29",
+    "vue": "^3.3.8",
+    "wasm-pack": "^0.12.1",
     "weak-napi": "^2.0.2",
     "yargs-parser": "^21.1.1"
   },
   "overrides": {
+    "axios": "^1.6.1",
     "semver": "^7.5.4"
   },
   "files": [
+    "dist/*.node",
     "dist/**/*.js",
     "dist/*.d.ts",
     "dist/bin/rollup",
     "dist/es/package.json"
   ],
   "engines": {
-    "node": ">=14.18.0",
+    "node": ">=18.0.0",
     "npm": ">=8.0.0"
   },
   "exports": {
