@@ -38,10 +43,11 @@
     "dist",
     "client.d.ts",
     "index.cjs",
+    "index.d.cts",
     "types"
   ],
   "engines": {
-    "node": "^14.18.0 || >=16.0.0"
+    "node": "^18.0.0 || >=20.0.0"
   },
   "repository": {
     "type": "git",
